Move onEncryptionKeyRequest to client/encryption
